completed: login screen, home screen, detailed screen
login screen connected to heroku to be able to use parse for user log in. Missing fields will not allow user to continue. Completed static version of homeview, with the use of the game superSmash bro. as an example. Completed static version of detailed view controller with the use of super smash bros. as an example